requirements.txt
misskey.py
websockets

python-dotenv
